+++ example/barco_event_master_example.dart
@@ -1,11 +1,9 @@
-import 'dart:io';
-
 import 'package:barco_event_master/barco_event_master.dart';
 
 Future<void> main() async {
-  final e2 = EventMaster(address: InternetAddress('192.168.1.100'));
-  final r = await e2.send(
-    ChangeAuxContentParams(auxId: 1, programSourceIndex: 13).toRequest(),
-  );
-  print(r.toJson());
+  final e2 = EventMaster(ipAddress: '37.101.8.90');
+
+  final r = await e2.listDestinations();
+
+  print(r.result?.response?.toJson());
 }