+++ lib/barco_event_master.dart
@@ -1,4 +1,6 @@
 library barco_event_master;
 
 export 'src/barco_event_master_base.dart';
-export 'src/models.dart';
+export 'src/models/event_master_response.dart';
+export 'src/models/params/params.dart';
+export 'src/models/response/response.dart';