+++ lib/src/barco_event_master_base.dart
@@ -1,26 +1,46 @@
 import 'dart:convert';
-import 'dart:io';
-import 'package:http/http.dart' as http;
+import 'dart:math';
 
-import 'package:barco_event_master/src/models/event_master_request.dart';
-import 'package:barco_event_master/src/models/event_master_response.dart';
+import 'package:barco_event_master/barco_event_master.dart';
+import 'package:http/http.dart' as http;
 
 class EventMaster {
   EventMaster({
-    required this.address,
+    required this.ipAddress,
   });
 
-  InternetAddress address;
+  final String ipAddress;
 
-  Uri get _uri => Uri.parse('http://${address.host}:9999');
+  Uri get _uri => Uri.parse('http://$ipAddress:9999');
 
-  Future<EventMasterResponse> send(EventMasterRequest request) async {
+  String get _randomId => Random().nextInt(99999).toString();
+
+  Future<EventMasterResponse<Map<String, dynamic>>> send({
+    required String method,
+    required Map<String, dynamic> params,
+    String? id,
+  }) async {
     final r = await http.post(
       _uri,
-      body: request.toJson(),
+      body: json.encode({
+        'method': method,
+        'params': params,
+        'id': id ?? _randomId,
+      }),
       encoding: Encoding.getByName('application/json'),
     );
 
     return EventMasterResponse.fromJson(r.body);
+  }
+}
+
+extension AddIfPresent<K, V> on Map<K, V> {
+  /// Adds a key/value pair as long as the value is not null or an empty String
+  void addIfPresent(K key, V value) {
+    if (value != null) {
+      if (value is! String || value.isNotEmpty) {
+        addAll({key: value});
+      }
+    }
   }
 }