+++ lib/src/models/params/change_aux_content.dart
@@ -1,47 +1,41 @@
-import 'dart:convert';
+import 'package:barco_event_master/barco_event_master.dart';
 
-import 'package:barco_event_master/src/models/params/base.dart';
+extension ChangeAuxContent on EventMaster {
+  /// Changes the source of an Aux destination
+  ///
+  /// Must specify: [auxId] OR [auxName] for destination
+  /// Must specifiy: [presetSerialNumber]
+  Future<EventMasterResponse<Map<String, dynamic>>> changeAuxContent({
+    /// Index of the aux destination
+    required int auxId,
 
-/// Changes the source of an Aux destination
-///
-/// Must specify: [auxId] OR [auxName] for destination
-/// Must specifiy: [presetSerialNumber]
-class ChangeAuxContentParams extends BaseEventMasterParam {
-  ChangeAuxContentParams({
-    required this.auxId,
-    this.auxName = '',
-    this.previewSourceIndex,
-    this.programSourceIndex,
-  });
+    /// Preset name
+    String auxName = '',
 
-  @override
-  String get method => 'changeAuxContent';
+    /// Source to be sent to Aux preview
+    int? previewSourceIndex,
 
-  /// Index of the aux destination
-  int auxId;
+    /// Source to be sent to Aux program
+    int? programSourceIndex,
 
-  /// Preset name
-  String auxName;
+    /// ID of a test pattern to be sent to the Aux
+    int? testPatternId,
 
-  /// Source to be sent to Aux preview
-  int? previewSourceIndex;
-
-  /// Source to be sent to Aux program
-  int? programSourceIndex;
-
-  /// ID of a test pattern to be sent to the Aux
-  int? testPatternId;
-
-  @override
-  Map<String, dynamic> toMap() {
-    return {
+    /// ID of the request used to match the incoming response (optional)
+    String? id,
+  }) async {
+    final params = <String, Object?>{
       'id': auxId,
     }
       ..addIfPresent('Name', auxName)
       ..addIfPresent('PvwLastSrcIndex', previewSourceIndex)
-      ..addIfPresent('PgmLastSrcIndex', programSourceIndex);
-  }
+      ..addIfPresent('PgmLastSrcIndex', programSourceIndex)
+      ..addIfPresent('TestPattern', testPatternId);
 
-  @override
-  String toJson() => json.encode(toMap());
+    return await send(
+      method: 'changeAuxContent',
+      params: params,
+      id: id,
+    );
+  }
 }