+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0
+
+- Rework requests so they return typesafe data objects
+- Remove dependency on dart:io so that JS compilation is supported
+- Add listDestinations request
+
 ## 1.0.0
 
 - Initial version.